// src/services/siswa/siswa.type.ts
import { TKelas } from "../kelas/kelas.type";
import { TKeluarga } from "../keluarga/keluarga.type";
import { TTahunAjaran } from "../tahun-ajaran/tahun-ajaran.type";

export type TSiswa = {
  id: number;
  nisn: string;
  nis: string;
  nik: string;
  nama: string;
  agama: string;
  tanggal_lahir: string;
  gol_darah?: null | string;
  type?: null | string;
  alamat: string;
  no_telp?: null | string;
  bahasa_sehari: string;
  ibu_kandung: string;
  kewarganegaraan: string;
  tanggal_masuk: string;
  lulusan_dari: string;
  jenis_kelamin: string;
  alamat_sekolah_asal: string;
  url_kia: string;
  url_kk: string;
  url_akta: string;
  avatar: string;
  telp_rumah: string;
  suku: string;
  tinggi_badan: string;
  berat_badan: string;
  lingkar_kepala: string;
  tempat_lahir: string;
  npsn_asal: string;
  tahun_lulus_asal: string;
  tinggal_bersama: string;
  anak_ke: string;
  jumlah_saudara: string;
  tanggal_keluar?: null | string;
  kelas?: TKelas[];
  keluarga: TKeluarga[];
  history_kelas: {
    id: number;
    siswa_id: number;
    tahun_ajaran_id: number;
    tahun_ajaran: TTahunAjaran;
    kelas: TKelas;
  }[];
  created_at: string;
  updated_at: string;
};

export type TPpdb = {
  id: number;
  nama_siswa: string;
  npsn_asal: string;
  kelas_id: string;
  tingkatan_id: string;
  tanggal_lahir: string;
  alamat: string;
  types: string;
  sekolah_asal: string;
  nama_orang_tua: string;
  no_hp_orang_tua: string;
  reason: string;
  know_from: string;
  biaya_pendaftaran: null | string;
  bukti_pembayaran: null | string;
  payment_method: null | string;
  token: string;
  status: string;
  data_siswa: any;
  data_ortu: any;
  is_past_to_test: boolean;
  payment_at: null | string;
  is_lunas: boolean;
  created_at: string;
  updated_at: string;
};